Exercise 12.2
int[] counts(int[] numbers,int size) {
...
}
that counts the number of negative values, the number of zeros and the
number of positive values in the array array of length size. The function
returns an array (say A) with 3 elements such that A[0] is the number of
negative values, A[1] is the number of zeros and A[2] is the number of positive
value.